airflow-docker/flaskapp/streamrequirements.txt
blinker==1.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-MySQL==1.5.2
fonttools==4.53.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpld3==0.5.10
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
py4j==0.10.9.7
PyMySQL==1.1.0
pyparsing==3.1.2
pyspark>=3.0.0, <3.6.0  # Loosened version constraint
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.2
cryptography